fix: split types out from usePerspective to prevent importing sdk in cypress
