+++ web/cypress/e2e/alerts/alerts_regression.cy.ts
@@ -1,4 +1,4 @@
-import { CustomerPerspectiveName } from '@/shared/hooks/usePerspective';
+import { CustomerPerspectiveName } from '@/shared/constants/perspective';
 import {
   testAlertsCorePlatformHeaderRegression,
   testAlertsRegression,
