+++ web/cypress/e2e/perses-dashboards/perses_dashboards_create.cy.ts
@@ -1,4 +1,4 @@
-import { CustomerPerspectiveName } from '@/shared/hooks/usePerspective';
+import { CustomerPerspectiveName } from '@/shared/constants/perspective';
 import { nav } from '../../views/nav';
 import { testCOOCreatePerses } from '../../support/perses/03.coo_create_perses_admin.cy';
 import { operatorAuthUtils } from '../../support/commands/auth-commands';
