+++ web/cypress/e2e/perses-dashboards/perses_dashboards_datasources.cy.ts
@@ -1,4 +1,4 @@
-import { CustomerPerspectiveName } from '@/shared/hooks/usePerspective';
+import { CustomerPerspectiveName } from '@/shared/constants/perspective';
 import {
   CLUSTER_LOGGING_OPERATOR,
   CLUSTER_MONITORING_OPERATOR,
