+++ web/cypress/e2e/perses-dashboards/perses_dashboards_import.cy.ts
@@ -1,4 +1,4 @@
-import { CustomerPerspectiveName } from '@/shared/hooks/usePerspective';
+import { CustomerPerspectiveName } from '@/shared/constants/perspective';
 import { nav } from '../../views/nav';
 import { testCOOImportPerses } from '../../support/perses/04.coo_import_perses_admin.cy';
 import { operatorAuthUtils } from '../../support/commands/auth-commands';
