+++ web/cypress/support/monitoring/04.reg_alerts_namespace.cy.ts
@@ -17,7 +17,7 @@ import {
   WatchdogAlert,
 } from '../../fixtures/monitoring/constants';
 import { FilterOUIAIDs } from '@/shared/constants/data-test';
-import type { CustomerPerspective } from '@/shared/hooks/usePerspective';
+import type { CustomerPerspective } from '@/shared/constants/perspective';
 
 export function testAlertsRegressionNamespace(perspectiveName: CustomerPerspective) {
   it(`${perspectiveName} perspective - Alerting > Alerts page - Filtering`, () => {
