+++ web/cypress/support/monitoring/05.reg_metrics_namespace_1.cy.ts
@@ -6,7 +6,7 @@ import {
   MetricsPageQueryInput,
   MetricsPageUnits,
 } from '../../fixtures/monitoring/constants';
-import type { CustomerPerspective } from '@/shared/hooks/usePerspective';
+import type { CustomerPerspective } from '@/shared/constants/perspective';
 
 export function testMetricsRegressionNamespace1(perspectiveName: CustomerPerspective) {
   it(`${perspectiveName} perspective - Metrics`, () => {
