+++ web/cypress/support/perses/01.coo_list_perses_admin.cy.ts
@@ -6,7 +6,7 @@ import { commonPages } from '../../views/common';
 import { listPersesDashboardsPage } from '../../views/perses-dashboards-list-dashboards';
 import { persesDashboardsPage } from '../../views/perses-dashboards';
 import { nav } from '../../views/nav';
-import type { CustomerPerspective } from '@/shared/hooks/usePerspective';
+import type { CustomerPerspective } from '@/shared/constants/perspective';
 
 export function testCOOListPerses(
   perspectiveName: CustomerPerspective,
