+++ web/cypress/support/perses/99.coo_rbac_perses_user5.cy.ts
@@ -12,7 +12,7 @@ import { persesDashboardsPanel } from '../../views/perses-dashboards-panel';
 import { persesDashboardsAddListPanelType } from '../../fixtures/perses/constants';
 import { persesImportDashboardsPage } from '../../views/perses-dashboards-import-dashboard';
 import { nav } from '../../views/nav';
-import type { CustomerPerspective } from '@/shared/hooks/usePerspective';
+import type { CustomerPerspective } from '@/shared/constants/perspective';
 
 let dashboardName = 'Testing Dashboard - UP ';
 const randomSuffix = Math.random().toString(5);
