+++ web/cypress/support/perses/99.coo_rbac_perses_user6.cy.ts
@@ -1,5 +1,5 @@
 import { listPersesDashboardsPage } from '../../views/perses-dashboards-list-dashboards';
-import type { CustomerPerspective } from '@/shared/hooks/usePerspective';
+import type { CustomerPerspective } from '@/shared/constants/perspective';
 
 /**
  * User6 has access to:
